The great proto3 migration
In order to ensure we are using libraries that are carefully audited
for their security properties on a regular basis, this CL moves from
the Square wire proto library to Google's official proto3 javalite
implementation.

Furthermore, to resolve the "spi depends on api" problem, the core
protos and data types have been moved to a new "protocol" module that
api and spi depend on (in addition to bbq).

This has resulted in widespread changes across the codebase, though
primarily to imports and protocol buffer interactions. This will
most likely break service provider implementations, though fixing
this mostly entails calling the method-form to access proto fields.
Sorry about this; we felt it was necessary to get this out of the
way before attempting to ratify an implementer's draft of OpenYOLO
Android.