Remove usage of response intents from retrieve BBQ flow (#132)
Fix for #131. In order to avoid the potential for code injection and request forgery attacks
from malicious providers, this change deprecates the use of intents as part of the retrieve
BBQ flow. Instead, providers now signal the availability of credentials via a simple boolean.
Optionally, they can also specify preloaded data that the client should send back to them
verbatim, as a replacement for the ability to specify intent extras on returned intents.