Increase timeout for providers, add new timeout status code (#136)
During testing with @nerdyverde we noticed that if the device is busy (such as
installing app updates) that the two second timeout is often insufficient, resulting
in failure to retrieve. This PR boosts the timeout to 4 seconds, and introduces
an explicit timeout error code to make it easier to understand when this happens.